Changed MatchDto id to String

+++ src/main/kotlin/com/bashkevich/tennisscorekeeperbackend/feature/match/websocket/MatchObserver.kt
@@ -15,7 +15,7 @@ object MatchObserver {
     }
 
     suspend fun notifyChange(match: MatchDto) {
-        val id = match.id
+        val id = match.id.toInt()
         val flow = matchFlows.getOrPut(id) {
             MutableSharedFlow(replay = 1)
         }
